ENH: Partially fix writing the best model output figure file
Partially fix writing the best model output figure file: deep-copy the
axes contents prior to providing those to the plot method.

The file at least contains the ground truth train/valid plots, and the
title indicates the best metric/epoch. However, the predictions and the
loss and metric values are still not plot.