ENH: Compute performance measurements for labels that are present
Compute performance measurements for labels that are present: use the
intersection of the sets instead of the union.

Otherwise, this makes such that performances computed by the
`seg_metrics` package for missing labels take values (which, for
distance-based measures such as the Hausdorff distance, are the maximum
allowed value for the data type, e.g. 1e+19) that bias any downstream
computation, and make the stats unusable.

For the missing labels, the performance measure values are now set
depending on the measure that is being considered, and assuming that the
ground truth labelmap contains all labels. This may be improved in
future PRs.

Use `np.inf` values for missing distance-based performance values: when
a label is missing (e.g. because it has not been predicted),
`seg_metrics` returns very large (probably the maximum value for the
data type) values for distance-based measures (e.g. Hausdorff distance,
mean surface distance, etc.). This makes it such that the stats computed
across participants are of little value (as the mean, std, etc. will
have very large values).

When an `np.inf` value is encountered, stats will be affected as well,
but the value can easily be identified and filtered in pandas to avoid
this effect.

For the rest of the values, the corresponding theoretical worst values
are used. Results are not affected by this change, since for
non-distance-based measures (e.g. dice, jaccard), the value that
`seg_metrics` was computing was the worst possible value.

Use a context manager that sets `mode.use_inf_as_na` to `True` when
computing the stats of the pandas df so that the `np.inf` values do not
make the stats unusable.