BUG: Fix the stats computation
Fix the stats computation: we are interested in computing first the
performance across labels (columns), and then computing the stats
across participants (rows). When using

```
df_subset.to_numpy().flatten()
```

all data is flattened and the stats are not the same: when there is no
NaN or np.inf, the mean will be the same, but not the rest of the stats
(count, std dev, min, max, percentiles); when there is NaN or np.inf,
all stats are different.

So, instead of flattening all values and using the context manager with
`mode.use_inf_as_na`, turn +/-np.inf values to NaNs first and compute
the mean across columns (`axis=1`) prior to computing the stats across
participants.

When calling describe on the entire dataframe, stats are computed on
each column separately. Thus, the above does not affect the code when
column-wise stats (i.e. label-wise) stats are computed (e.g.
`compute_performance.py` and `aggregate_performance_folds.py`)
(provided that we use thecontext manager with `mode.use_inf_as_na`).

Add additional checks to warn the user when all values for a given label
group (e.g. the fastigial) are NaNs (np.inf values are translated into
NaNs previously) (i.e. missing data: e.g. the fastigial was missed by
the model in the prediction).

Move the checks that previously existed in the anova and pairwise test
scripts about the dice measures not containing NaN or np.inf values
to the stat preparation utils module, refactoring them into the above
checks to warn users about missing data.