ENH: Allow labelmaps and label visibility in Slicer screenshot script
Allow loading labelmaps for capturing data view screenshots in 3D
Slicer, including views that have been zoomed/panned.

Also, allow to change the visibility of the labels so that only those of
interest are visible when capturing the screenshot.

Add the shell script that uses these features and which was used for the
ISMRM abstract.