ENH: Extend pairwise test script to subsampling experiments
Extend the pairwise test script to subsampling experiments.

Specifically, filter data from participants that are not present across
all experiments. Some participants did not have the required number of
diffusion sensitizing directions, and thus were not included in the
subsampling experiments. An assertion exists in the script to ensure
that all participants are present across data frames.

Fixes:
```
  File "/home/jhlegarreta/src/dmriseg/scripts/compute_pairwise_test.py", line 97, in main
    assert not any(
AssertionError
```

Add the accompanying shell script.