BUG: Handle appropriately optional labels argument in pairwise test
Handle appropriately optional labels argument in pairwise test
script: only attempt to create the list of labels if these were
specified; use `None` otherwise.

The columns of interest in the pairwise test data preparation function
are optional. If `None`, all `SUIT` labels are used for the analysis.

Fixes:
```
  File "scripts/compute_pairwise_test.py", line 111, in main
    columns_of_interest=list(map(str, args.labels)),
TypeError: 'NoneType' object is not iterable
```